+++ src/toast/schedule_sim_ground.py
@@ -249,7 +249,7 @@ class Patch(object):
             az2 = unwind_angle(az1, azs[j])
             el2 = els[j]
             azmean = 0.5 * (az1 + az2)
-            az0 = unwind_angle(azmean, np.float(obj.az), np.pi)
+            az0 = unwind_angle(azmean, float(obj.az), np.pi)
             if (az1 - az0) * (az2 - az0) > 0:
                 # the constant meridian is not between the two corners
                 continue
@@ -278,7 +278,7 @@ class Patch(object):
                 els_cross[:i] += 2 * np.pi
                 els_cross = np.sort(els_cross)
         el_mean = np.mean(els_cross)
-        el0 = unwind_angle(el_mean, np.float(obj.alt))
+        el0 = unwind_angle(el_mean, float(obj.alt))
 
         ncross = np.sum(els_cross > el0)
 
@@ -905,7 +905,7 @@ def check_sso(observer, az1, az2, el, sso, angle, el_min, tstart, tstop):
     """
     if az2 < az1:
         az2 += 360
-    naz = max(3, np.int(0.25 * (az2 - az1) * np.cos(np.radians(el))))
+    naz = max(3, int(0.25 * (az2 - az1) * np.cos(np.radians(el))))
     quats = []
     for az in np.linspace(az1, az2, naz):
         quats.append(from_angles(az % 360, el))
@@ -1606,7 +1606,7 @@ def add_scan(
     log = Logger.get()
     ces_time = tstop - tstart
     if ces_time > args.ces_max_time_s:  # and not args.pole_mode:
-        nsub = np.int(np.ceil(ces_time / args.ces_max_time_s))
+        nsub = int(np.ceil(ces_time / args.ces_max_time_s))
         ces_time /= nsub
     aztimes = np.array(aztimes)
     azmins = np.array(azmins)
@@ -2055,7 +2055,7 @@ def build_schedule(args, start_timestamp, stop_timestamp, patches, observer, sun
         el_min = 90
         el_max = 0
         for el in args.elevations_deg.split(","):
-            el = np.float(el)
+            el = float(el)
             el_min = min(el * 0.9, el_min)
             el_max = max(el * 1.1, el_max)
     el_min *= degree
@@ -2286,14 +2286,14 @@ def parse_args(opts=None):
         "--site-alt",
         required=False,
         default=100,
-        type=np.float,
+        type=float,
         help="Observing site altitude [meters]",
     )
     parser.add_argument(
         "--scan-margin",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Random fractional margin [0..1] added to the "
         "scans to smooth out edge effects",
     )
@@ -2301,49 +2301,49 @@ def parse_args(opts=None):
         "--ra-period",
         required=False,
         default=10,
-        type=np.int,
+        type=int,
         help="Period of patch position oscillations in RA [visits]",
     )
     parser.add_argument(
         "--ra-amplitude-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Amplitude of patch position oscillations in RA [deg]",
     )
     parser.add_argument(
         "--dec-period",
         required=False,
         default=10,
-        type=np.int,
+        type=int,
         help="Period of patch position oscillations in DEC [visits]",
     )
     parser.add_argument(
         "--dec-amplitude-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Amplitude of patch position oscillations in DEC [deg]",
     )
     parser.add_argument(
         "--elevation-penalty-limit",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Assign a penalty to observing elevations below this limit [degrees]",
     )
     parser.add_argument(
         "--elevation-penalty-power",
         required=False,
         default=2,
-        type=np.float,
+        type=float,
         help="Power in the elevation penalty function [> 0]",
     )
     parser.add_argument(
         "--elevation-change-limit-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Assign a penalty to changes in elevation larger than this limit [degrees].  "
         "See --elevation-change-penalty and --elevation-change-time-s",
     )
@@ -2351,7 +2351,7 @@ def parse_args(opts=None):
         "--elevation-change-penalty",
         required=False,
         default=1,
-        type=np.float,
+        type=float,
         help="Multiplicative elevation change penalty triggered by "
         "--elevation-change-limit-deg",
     )
@@ -2359,7 +2359,7 @@ def parse_args(opts=None):
         "--elevation-change-time-s",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Time it takes for the telescope to stabilize after a change in observing "
         "elevation [seconds].  Triggered by --elevation-change-limit-deg",
     )
@@ -2403,21 +2403,21 @@ def parse_args(opts=None):
         "--el-min-deg",
         required=False,
         default=30,
-        type=np.float,
+        type=float,
         help="Minimum elevation for a CES",
     )
     parser.add_argument(
         "--el-max-deg",
         required=False,
         default=80,
-        type=np.float,
+        type=float,
         help="Maximum elevation for a CES",
     )
     parser.add_argument(
         "--el-step-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Optional step to apply to minimum elevation",
     )
     parser.add_argument(
@@ -2431,70 +2431,70 @@ def parse_args(opts=None):
         "--fp-radius-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Focal plane radius [deg]",
     )
     parser.add_argument(
         "--sun-avoidance-angle-deg",
         required=False,
         default=30,
-        type=np.float,
+        type=float,
         help="Minimum distance between the Sun and the bore sight [deg]",
     )
     parser.add_argument(
         "--sun-avoidance-altitude-deg",
         required=False,
         default=-18,
-        type=np.float,
+        type=float,
         help="Minimum altitude to apply Solar avoidance [deg]",
     )
     parser.add_argument(
         "--moon-avoidance-angle-deg",
         required=False,
         default=20,
-        type=np.float,
+        type=float,
         help="Minimum distance between the Moon and the bore sight [deg]",
     )
     parser.add_argument(
         "--moon-avoidance-altitude-deg",
         required=False,
         default=-18,
-        type=np.float,
+        type=float,
         help="Minimum altitude to apply Lunar avoidance [deg]",
     )
     parser.add_argument(
         "--sun-el-max-deg",
         required=False,
         default=90,
-        type=np.float,
+        type=float,
         help="Maximum allowed sun elevation [deg]",
     )
     parser.add_argument(
         "--boresight-angle-step-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Boresight rotation step size [deg]",
     )
     parser.add_argument(
         "--boresight-angle-min-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Boresight rotation angle minimum [deg]",
     )
     parser.add_argument(
         "--boresight-angle-max-deg",
         required=False,
         default=360,
-        type=np.float,
+        type=float,
         help="Boresight rotation angle maximum [deg]",
     )
     parser.add_argument(
         "--boresight-angle-time-min",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Boresight rotation step interval [minutes]",
     )
     parser.add_argument(
@@ -2516,13 +2516,13 @@ def parse_args(opts=None):
     parser.add_argument(
         "--operational-days",
         required=False,
-        type=np.int,
+        type=int,
         help="Number of operational days to schedule (empty days do not count)",
     )
     parser.add_argument(
         "--timezone",
         required=False,
-        type=np.int,
+        type=int,
         default=0,
         help="Offset to apply to MJD to separate operational days [hours]",
     )
@@ -2530,21 +2530,21 @@ def parse_args(opts=None):
         "--gap-s",
         required=False,
         default=100,
-        type=np.float,
+        type=float,
         help="Gap between CES:es [seconds]",
     )
     parser.add_argument(
         "--gap-small-s",
         required=False,
         default=10,
-        type=np.float,
+        type=float,
         help="Gap between split CES:es [seconds]",
     )
     parser.add_argument(
         "--time-step-s",
         required=False,
         default=600,
-        type=np.float,
+        type=float,
         help="Time step after failed target acquisition [seconds]",
     )
     parser.add_argument(
@@ -2558,7 +2558,7 @@ def parse_args(opts=None):
         "--ces-max-time-s",
         required=False,
         default=900,
-        type=np.float,
+        type=float,
         help="Maximum length of a CES [seconds]",
     )
     parser.add_argument(
@@ -2576,13 +2576,13 @@ def parse_args(opts=None):
     parser.add_argument(
         "--pol-min",
         required=False,
-        type=np.float,
+        type=float,
         help="Lower plotting range for polarization map",
     )
     parser.add_argument(
         "--pol-max",
         required=False,
-        type=np.float,
+        type=float,
         help="Upper plotting range for polarization map",
     )
     parser.add_argument(
@@ -2603,14 +2603,14 @@ def parse_args(opts=None):
         "--pole-el-step-deg",
         required=False,
         default=0.25,
-        type=np.float,
+        type=float,
         help="Elevation step in pole scheduling mode [deg]",
     )
     parser.add_argument(
         "--pole-ces-time-s",
         required=False,
         default=3000,
-        type=np.float,
+        type=float,
         help="Time to scan at constant elevation in pole mode",
     )
     parser.add_argument(
@@ -2620,14 +2620,14 @@ def parse_args(opts=None):
         "--boresight-offset-el-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Optional offset added to every observing elevation",
     )
     parser.add_argument(
         "--boresight-offset-az-deg",
         required=False,
         default=0,
-        type=np.float,
+        type=float,
         help="Optional offset added to every observing azimuth",
     )
     parser.add_argument(
@@ -2662,35 +2662,35 @@ def parse_args(opts=None):
         "--pole-raster-el-step-deg",
         required=False,
         default=1 / 60,
-        type=np.float,
+        type=float,
         help="Elevation step in pole raster scheduling mode [deg]",
     )
     parser.add_argument(
         "--az-rate-sky-deg",
         required=False,
         default=1.0,
-        type=np.float,
+        type=float,
         help="Azimuthal rate in pole raster scheduling mode [deg]",
     )
     parser.add_argument(
         "--az-accel-mount-deg",
         required=False,
         default=1.0,
-        type=np.float,
+        type=float,
         help="Azimuthal accleration in pole raster scheduling mode [deg]",
     )
     parser.add_argument(
         "--el-rate-deg",
         required=False,
         default=1.0,
-        type=np.float,
+        type=float,
         help="Elevation rate in pole raster scheduling mode [deg]",
     )
     parser.add_argument(
         "--el-accel-deg",
         required=False,
         default=1.0,
-        type=np.float,
+        type=float,
         help="Elevation accleration in pole raster scheduling mode [deg]",
     )
 
@@ -3051,8 +3051,8 @@ def parse_patches(args, observer, sun, moon, start_timestamp, stop_timestamp):
         moon_avoidance_color = "gray"
         alpha = 0.5
         avoidance_alpha = 0.01
-        sun_step = np.int(86400 * 1)
-        moon_step = np.int(86400 * 0.1)
+        sun_step = int(86400 * 1)
+        moon_step = int(86400 * 0.1)
         for iplot, coord in enumerate("CEG"):
             scoord = {"C": "Equatorial", "E": "Ecliptic", "G": "Galactic"}[coord]
             title = scoord  # + ' patch locations'
@@ -3102,7 +3102,7 @@ def parse_patches(args, observer, sun, moon, start_timestamp, stop_timestamp):
                     moon_lw,
                 ),
             ]:
-                for t in range(np.int(start_timestamp), np.int(stop_timestamp), step):
+                for t in range(int(start_timestamp), int(stop_timestamp), step):
                     observer.date = to_DJD(t)
                     sso.compute(observer)
                     lon.append(np.degrees(sso.a_ra))
